DF-11859 fix for review
Former-commit-id: bfe9f17994ec095874775028e831ca89fe1473af

+++ Sources/Internal/UI/Private/Android/MovieViewControlAndroid.h
@@ -93,7 +93,7 @@ public:
     bool IsPlaying() const;
 
 private:
-    uint32 id;
+    uintptr_t id;
 
     JNI::JavaClass jniMovieViewControl;
     Function<void(jint, jfloat, jfloat, jfloat, jfloat)> initialize;
